+++ src/TowerDefense/modele/tourelle/TourelleRoche.java
@@ -8,7 +8,7 @@ import TowerDefense.modele.tourelle.Tourelle;
 
 public class TourelleRoche extends Tourelle {
 
-	public TourelleRoche(int x, int y, Terrain terrain,Jeu jeu) {
+	public TourelleRoche(double x, double y, Terrain terrain,Jeu jeu) {
 		super(x, y, terrain, jeu);
 	}
